+++ test/OpenFeature.Extensions.Hosting.Tests/HostingTest.cs
@@ -1,8 +1,7 @@
-using System;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
+using OpenFeature.Extensions.Hosting.Tests.TestingModels;
 using OpenFeature.Model;
 using Xunit;
 
@@ -24,7 +23,8 @@ public sealed class HostingTest
 #pragma warning restore xUnit1030
 
         Assert.Equal(Api.Instance, app.Services.GetRequiredService<Api>());
-        Assert.Equal(Api.Instance.GetProviderMetadata().Name, app.Services.GetRequiredService<IFeatureClient>().GetMetadata().Name);
+        Assert.Equal(Api.Instance.GetProviderMetadata().Name,
+            app.Services.GetRequiredService<IFeatureClient>().GetMetadata().Name);
 
         Assert.Empty(Api.Instance.GetContext().AsDictionary());
         Assert.Empty(app.Services.GetRequiredService<EvaluationContextBuilder>().Build().AsDictionary());
@@ -69,48 +69,5 @@ public sealed class HostingTest
 #pragma warning disable xUnit1030
         await app.StopAsync().ConfigureAwait(false);
 #pragma warning restore xUnit1030
-    }
-}
-
-sealed class SomeFeatureProvider : FeatureProvider
-{
-    public const string Name = "some_feature_provider";
-
-    public override Metadata GetMetadata() => new(Name);
-
-    public override Task<ResolutionDetails<bool>> ResolveBooleanValue(string flagKey, bool defaultValue, EvaluationContext? context = null)
-        => Task.FromResult(new ResolutionDetails<bool>(flagKey, defaultValue));
-
-    public override Task<ResolutionDetails<string>> ResolveStringValue(string flagKey, string defaultValue, EvaluationContext? context = null)
-        => Task.FromResult(new ResolutionDetails<string>(flagKey, defaultValue));
-
-    public override Task<ResolutionDetails<int>> ResolveIntegerValue(string flagKey, int defaultValue, EvaluationContext? context = null)
-        => Task.FromResult(new ResolutionDetails<int>(flagKey, defaultValue));
-
-    public override Task<ResolutionDetails<double>> ResolveDoubleValue(string flagKey, double defaultValue, EvaluationContext? context = null)
-        => Task.FromResult(new ResolutionDetails<double>(flagKey, defaultValue));
-
-    public override Task<ResolutionDetails<Value>> ResolveStructureValue(string flagKey, Value defaultValue, EvaluationContext? context = null)
-        => Task.FromResult(new ResolutionDetails<Value>(flagKey, defaultValue));
-}
-
-public static class SomeFeatureProviderExtensions
-{
-    public static OpenFeatureBuilder AddSomeFeatureProvider(this OpenFeatureBuilder builder)
-    {
-        if (builder == null)
-        {
-            throw new ArgumentNullException(nameof(builder));
-        }
-
-        builder.ProvidersServiceCollection.TryAddEnumerable(ServiceDescriptor.Singleton<FeatureProvider, SomeFeatureProvider>());
-        builder.TryAddOpenFeatureClient(SomeFeatureProvider.Name);
-
-        return builder;
-    }
-
-    public static void AddSomeFeatureProvider(this OpenFeatureBuilder builder, Action<OpenFeatureBuilder> configure)
-    {
-        throw new NotImplementedException();
     }
 }
